+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Debounce unread message counter in timeline to avoid flickering
-- Focus message input text field on non-mobile platforms when starting a reply
+- Focus message input text field when starting a reply
 
 ### Deprecated